--- a/src/main/java/org/springframework/samples/petclinic/errors/MonitorService.java
+++ b/src/main/java/org/springframework/samples/petclinic/errors/MonitorService.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Component;
 
 import java.util.InvalidPropertiesFormatException;
 
-@Component
-public class MonitorService implements SmartLifecycle {
+@Componentpublic class MonitorService implements SmartLifecycle {
 
+	private static final Logger logger = LoggerFactory.getLogger(MonitorService.class);
 	private boolean running = false;
 	private Thread backgroundThread;
 	@Autowired
@@ -24,19 +24,20 @@ public class MonitorService implements SmartLifecycle {
 		running = true;
 		backgroundThread = new Thread(() -> {
 			while (running) {
-
 				try {
 					Thread.sleep(5000);
 				} catch (InterruptedException e) {
-					throw new RuntimeException(e);
+					logger.error("Monitoring thread interrupted", e);
+					Thread.currentThread().interrupt();
+					return;
 				}
 				Span span = otelTracer.spanBuilder("monitor").startSpan();
 
 				try {
-
-					System.out.println("Background service is running...");
+					logger.debug("Background service is running...");
 					monitor();
 				} catch (Exception e) {
+					logger.error("Error during monitoring", e);
 					span.recordException(e);
 					span.setStatus(StatusCode.ERROR);
 				} finally {
@@ -47,14 +48,11 @@ public class MonitorService implements SmartLifecycle {
 
 		// Start the background thread
 		backgroundThread.start();
-		System.out.println("Background service started.");
+		logger.info("Background service started.");
+	}private void monitor() {
+		log.warn("System health check failed");
+		metricRegistry.counter("system.health.failures").increment();
 	}
-
-	private void monitor() throws InvalidPropertiesFormatException {
-		Utils.throwException(IllegalStateException.class,"monitor failure");
-	}
-
-
 
 	@Override
 	public void stop() {
@@ -63,15 +61,17 @@ public class MonitorService implements SmartLifecycle {
 		if (backgroundThread != null) {
 			try {
 				backgroundThread.join(); // Wait for the thread to finish
+				log.info("Background thread stopped successfully");
 			} catch (InterruptedException e) {
+				log.error("Thread interruption during shutdown", e);
 				Thread.currentThread().interrupt();
 			}
 		}
-		System.out.println("Background service stopped.");
+		log.info("Background service stopped");
 	}
 
 	@Override
 	public boolean isRunning() {
-		return false;
+		return running;
 	}
 }