This PR addresses two main issues:

1. Welcome Page Performance Optimization:
- Enabled template caching
- Configured resource chain strategy
- Added content versioning
- Implemented async loading for non-critical scripts
- Added lazy loading for images

2. MonitorService Improvements:
- Removed test exception
- Implemented proper system health monitoring
- Added appropriate logging
- Added system metrics collection

Related PRs: #313, #319, #317, #323

Performance Impact:
- Expected reduction in welcome page rendering time
- Current P50: 396.21ms → Expected: ~150ms
- Current P95: 413.19ms → Expected: ~200ms

Testing:
- Verified template caching
- Confirmed resource optimization
- Tested monitoring functionality
- Validated logging output